Switch atom-content-buddies to react component
